python-generator/requirements.txt
fastapi
uvicorn
google-genai
pillow
pillow-heif
requests
python-multipart
